connect and disconnect should return undefined
this because currently, we don't take a callback for connect, but
we should, because it should be possible to monitor the status of
the connect attmept, and if any error occurs, it's important to
handle it, rather than letting it fail silently.  If this method
is going to take a callback, then it should also be able to return
a promise using the promise library.  So to avoid an incompatible
API between the small callback library and the high level promise
library, have the callback library return something useless.
disconnect, for symmetry with connect.