KiteApi: wrap each method with a transformer method
This removes the need for munging callbacks etc. Instead of us rewriting the protocol, we are simply wrapping each function with a one that can understand the way we are sending messages, and transforms this dnode-request into our function calls by simply sending arguments in the correct order.

This also adds the request info as the last argument to the method handlers (see #63)

Signed-off-by: Umut Sirin <umutsirin1@gmail.com>